requirements/base.txt
# base.txt (Common dependencies for both dev and prod)

# Python 3.11

Django==5.1.5
djangorestframework==3.15.2
# djangorestframework-api-key==3.0.0
python-decouple==3.8
django-crontab==0.7.1
Pillow==10.2.0
# mysqlclient==2.2.0
django-cors-headers==4.3.1
django-safedelete==1.4.1
djangorestframework_simplejwt==5.4.0
# for postgres
psycopg2-binary==2.9.10
django-currentuser==0.9.0
# misc
celery==5.5.3
redis==6.2.0
django_celery_results==2.6.0
django-celery-beat==2.8.1

# for sanitizing rich text
nh3==0.3.0